requirements.txt
streamlit
google-genai
python-docx
PyMuPDF
